Improved server logic to handle exception and throw it out.

+++ src/main/java/com/group/eBookManagementSystem/repository/BookRepository.java
@@ -6,12 +6,11 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Repository
-public interface BookRepository extends CrudRepository<Book, Long> {
+public interface BookRepository extends CrudRepository<Book, Integer> {
 
     Book findBookById(Integer id);
 
     @Transactional
     void deleteBookById(Integer id);
-//    Customer findCustomerByFirstName(String firstName);
 
 }
